xilpm: versal: server: Fix MISRA 10.3 violation
Fix below MISRA-C violation:
- misra_c_2012_rule_10_3_violation:
	The value of an expression shall not be assigned to an object
	with a narrower essential type or of a different essential
	type category.


Acked-for-series: Jolly Shah <jolly.shah@xilinx.com>

+++ lib/sw_services/xilpm/src/versal/server/xpm_notifier.c
@@ -208,10 +208,10 @@ void XPmNotifier_Event(const u32 NodeId, const u32 Event)
 	Payload[2] = Event;
 
 	switch (NODECLASS(NodeId)) {
-		case XPM_NODECLASS_EVENT:
+		case (u32)XPM_NODECLASS_EVENT:
 			Payload[3] = 0U;
 			break;
-		case XPM_NODECLASS_DEVICE:
+		case (u32)XPM_NODECLASS_DEVICE:
 			Device = XPmDevice_GetById(NodeId);
 			if (NULL == Device) {
 				goto done;
